Add abstract EMsoft MP reader, use this in EBSD MP reader
